requirements.txt
aiogram==2.10.1
psycopg2-binary==2.8.6